backend/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0  # если используете PostgreSQL
drf-spectacular>=0.27.0
django-filter>=23.0
django-cors-headers>=4.3.0